# UP-4656:  Improve performance by caching the results of GroupMemberIm…