--- a/uportal-war/src/main/java/org/jasig/portal/api/permissions/ApiPermissionsService.java
+++ b/uportal-war/src/main/java/org/jasig/portal/api/permissions/ApiPermissionsService.java
@@ -82,7 +82,7 @@ public class ApiPermissionsService implements PermissionsService {
         
         if (includeInherited) {
             IGroupMember member = GroupService.getGroupMember(authP.getKey(), authP.getType());
-            for (@SuppressWarnings("unchecked") Iterator<IEntityGroup> iter = member.getAllContainingGroups(); iter.hasNext();) {
+            for (@SuppressWarnings("unchecked") Iterator<IEntityGroup> iter = member.getAncestorGroups(); iter.hasNext();) {
                 IEntityGroup parent = iter.next();
 
                 IAuthorizationPrincipal parentPrincipal = this.authorizationService.newPrincipal(parent);