--- a/uportal-war/src/main/java/org/jasig/portal/groups/IGroupMember.java
+++ b/uportal-war/src/main/java/org/jasig/portal/groups/IGroupMember.java
@@ -44,7 +44,6 @@ import org.jasig.portal.IBasicEntity;
  * that duplicates returned from "deep" methods can  be recognized.
  *
  * @author Dan Ellentuck
- * @version $Revision$
  */
 public interface IGroupMember extends IBasicEntity {
 /**
@@ -72,7 +71,7 @@ public boolean equals(Object o);
  *
  * @return java.util.Iterator
  */
-public Iterator getAllContainingGroups() throws GroupsException;
+public Iterator getAncestorGroups() throws GroupsException;
 /**
  * Returns an <code>Iterator</code> over the <code>Set</code> of this
  * <code>IGroupMember's</code> recursively-retrieved members that are
@@ -90,7 +89,7 @@ public Iterator getAllMembers() throws GroupsException;
  * Returns an <code>Iterator</code> over this <code>IGroupMember's</code> parent groups.
  * @return java.util.Iterator
  */
-public Iterator getContainingGroups() throws GroupsException;
+public Iterator getParentGroups() throws GroupsException;
 /**
  * Returns an <code>Iterator</code> over this <code>IGroupMember's</code>
  * members that are <code>IEntities</code>.