… uportal-war/src/main/java/org/jasig/portal/groups/RDBMEntityGroupStore.java
… uportal-war/src/main/java/org/jasig/portal/groups/filesystem/FileSystemGroupStore.java
… uportal-war/src/main/java/org/jasig/portal/groups/ldap/LDAPGroupStore.java
… uportal-war/src/main/java/org/jasig/portal/groups/pags/PersonAttributesGroupStore.java
… uportal-war/src/main/java/org/jasig/portal/groups/pags/dao/EntityPersonAttributesGroupStore.java
… uportal-war/src/main/java/org/jasig/portal/groups/smartldap/SmartLdapGroupStore.java
… uportal-war/src/main/java/org/jasig/portal/io/xml/portlet/PortletDefinitionImporterExporter.java
… uportal-war/src/main/java/org/jasig/portal/portlet/registry/PortletCategoryRegistryImpl.java
… uportal-war/src/main/java/org/jasig/portal/portlets/account/UserAccountHelper.java
… uportal-war/src/main/java/org/jasig/portal/portlets/groupadmin/GroupAdministrationHelper.java
… uportal-war/src/main/java/org/jasig/portal/rest/permissions/PermissionsRESTController.java
… uportal-war/src/main/java/org/jasig/portal/security/provider/AnyUnblockedGrantPermissionPolicy.java
… uportal-war/src/main/java/org/jasig/portal/security/provider/AuthorizationImpl.java
… uportal-war/src/main/java/org/jasig/portal/security/remoting/PermissionAssignmentMapController.java
… uportal-war/src/main/java/org/jasig/portal/services/GroupService.java
… uportal-war/src/main/java/org/jasig/portal/services/entityproperties/ContainingGroupsFinder.java
… uportal-war/src/main/java/org/jasig/portal/utils/threading/DoubleCheckedCreator.java
… uportal-war/src/main/resources/properties/ehcache.xml
… uportal-war/src/test/groovy/org/jasig/portal/io/xml/portlet/PortletDefinitionImporterExporterTest.groovy
… uportal-war/src/test/java/org/jasig/portal/groups/GroupsTester.java