…pl.getContainingGroups()

https://issues.jasig.org/browse/UP-4656

## Microbenchmarking

**WARNING:  These numbers reflect an early point in this PR and are not representative of the change.**

I'm using a server DEV deployment with MySQL and a lot of data (portlets & groups).

### Sample Requests to /Login Before (this patch)
```
TRACE [ajp-nio-127.0.0.1-8009-exec-8-andrewwills@district87.illinicloud.org] o.j.p.s.s.p.PortalPreAuthenticatedProcessingFilter 2016-04-06 18:22:16,241 - FINISHED PortalPreAuthenticatedProcessingFilter.doFilter() for ec7498d3-bfdb-44fb-a3c4-967b73530c59 in 5604ms #milestone
TRACE [ajp-nio-127.0.0.1-8009-exec-5-andrewwills@district87.illinicloud.org] o.j.p.s.s.p.PortalPreAuthenticatedProcessingFilter 2016-04-06 18:23:47,951 - FINISHED PortalPreAuthenticatedProcessingFilter.doFilter() for bb85c2bc-0792-486f-930b-53e03f9be1db in 3897ms #milestone
```

### Sample Requests to /Login After (this patch)
```
TRACE [ajp-nio-127.0.0.1-8009-exec-4-guest] o.j.p.s.s.p.PortalPreAuthenticatedProcessingFilter 2016-04-08 17:29:10,365 - FINISHED [8aa27d74-72fa-48ee-b69e-223058cb66c4] for URI=/uPortal/Login in 2677ms #milestone
TRACE [ajp-nio-127.0.0.1-8009-exec-2-andrewwills@district87.illinicloud.org] o.j.p.s.s.p.PortalPreAuthenticatedProcessingFilter 2016-04-08 17:30:09,705 - FINISHED [9d1b3b51-94cf-4574-be24-14af81c33e2b] for URI=/uPortal/Login in 3648ms #milestone
TRACE [ajp-nio-127.0.0.1-8009-exec-8-andrewwills@district87.illinicloud.org] o.j.p.s.s.p.PortalPreAuthenticatedProcessingFilter 2016-04-08 17:30:52,665 - FINISHED [66322f27-b62b-4480-a740-ccd7b5a90106] for URI=/uPortal/Login in 1919ms #milestone
TRACE [ajp-nio-127.0.0.1-8009-exec-5-andrewwills@district87.illinicloud.org] o.j.p.s.s.p.PortalPreAuthenticatedProcessingFilter 2016-04-08 17:31:29,558 - FINISHED [d8ee2d3d-4f78-47d1-a3be-50383a33c96f] for URI=/uPortal/Login in 2355ms #milestone
TRACE [ajp-nio-127.0.0.1-8009-exec-9-idda@district87.illinicloud.org] o.j.p.s.s.p.PortalPreAuthenticatedProcessingFilter 2016-04-08 17:33:18,148 - FINISHED [53691b1c-a6ba-475e-87f8-a886cfeee056] for URI=/uPortal/Login in 2354ms #milestone
TRACE [ajp-nio-127.0.0.1-8009-exec-8-andrewwills@district87.illinicloud.org] o.j.p.s.s.p.PortalPreAuthenticatedProcessingFilter 2016-04-08 17:35:19,512 - FINISHED [7db5278e-44fd-4e2e-b332-ad686ff68af0] for URI=/uPortal/Login in 8498ms #milestone
TRACE [ajp-nio-127.0.0.1-8009-exec-2-andrewwills@district87.illinicloud.org] o.j.p.s.s.p.PortalPreAuthenticatedProcessingFilter 2016-04-08 17:35:40,193 - FINISHED [59add746-4d6d-4d9b-bdd5-7c5187b152b4] for URI=/uPortal/Login in 3724ms #milestone
TRACE [ajp-nio-127.0.0.1-8009-exec-6-andrewwills@district87.illinicloud.org] o.j.p.s.s.p.PortalPreAuthenticatedProcessingFilter 2016-04-08 17:36:32,918 - FINISHED [cad6f345-2eec-42ff-bc07-33ab1e5e590b] for URI=/uPortal/Login in 7ms #milestone
TRACE [ajp-nio-127.0.0.1-8009-exec-5-andrewwills@district87.illinicloud.org] o.j.p.s.s.p.PortalPreAuthenticatedProcessingFilter 2016-04-08 17:36:43,617 - FINISHED [ba8769c0-6fa6-4533-9f26-e013a0c4641a] for URI=/uPortal/Login in 3000ms #milestone
TRACE [ajp-nio-127.0.0.1-8009-exec-10-andrewwills@district87.illinicloud.org] o.j.p.s.s.p.PortalPreAuthenticatedProcessingFilter 2016-04-08 17:37:08,668 - FINISHED [4d965de3-152d-4069-b5f0-207f29ac3f24] for URI=/uPortal/Login in 2976ms #milestone
```