@jameswennmacher 2016-04-08
Very good and promising start.  I'm looking forward to seeing the microbenchmarks of the improvement.