@drewwills 2016-04-09
## Update

This PR covers group membership evaluations going "up" the chain -- "who are my parents and ancestors?"

There is an equal issue (maybe worse?) going "down" the chain -- "who are my children and descendants?"