@drewwills 2016-04-11
## Update

Caching for going "down" the chain -- "who are my children and descendants?" -- has been added to this PR in the same manner as earlier.