@drewwills 2016-04-11
## Update

I don't have anything else planned for this PR.  It's ready for final comments and merge.