@acolebourne 2016-04-15
The branch in including the last commit built and runs for me.

Under load I'm seeing average response times of 5 seconds on marketplace. Logins 6 seconds average. Also the request throughput on Tomcat is up 5x compared to this branch -4 commits. Error rate is down from 1.5% to 0.75% - I attribute almost all of these to portlet apps.